[MOD-17813] Extract FT._LIST into src/coord/index_list_command.c
Moves IndexListInternal, IndexListCommandHandler, the fanout reducer
(IndexListClusterStateReducer) and everything private to them -
replyClusterStateEntry, the FT_LIST_CS_* keys, ClusterStateGates,
ClusterStateFingerprint, ClusterStateIndexInfo, ClusterStateReports,
ClusterStateVerdict - out of module.c and into a new
src/coord/index_list_command.c/.h pair. module.c was 5600+ lines before
this; the split mirrors src/coord/info_command.c/.h, which already holds
FT.INFO's InfoReplyReducer for the same reason. Requested twice by the PR
reviewer.

The extracted code called three module.c statics and read one module.c
global. SearchCluster_Ready, cannotBlockCtx and ReplyBlockDeny lose their
"static inline" and gain a declaration in module.h, becoming ordinary
externally-linked functions - nothing else about them changes. NumShards
is read through the existing GetNumShards_UnSafe() getter rather than
adding a new extern for the global itself, keeping the one accessor as
the only way to reach it from outside module.c.

Only code motion: no reply shape, classification rule or reducer logic
changes. Also settles the moved file's shape: the FT_LIST_CS_* keys move
up next to the includes instead of sitting mid-file, and the MRReply
pointers ClusterStateReports_AddShard only reads are now const.
