+++ src/coord/index_list_command.h
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+*/
+
+#pragma once
+
+#include "redismodule.h"
+
+int IndexListInternal(RedisModuleCtx *ctx, RedisModuleString **argv, int argc);
+
+int IndexListCommandHandler(RedisModuleCtx *ctx, RedisModuleString **argv, int argc);
