+++ src/module.c
@@ -64,6 +64,7 @@
 #include "coord/dist_profile.h"
 #include "coord/cluster_spell_check.h"
 #include "coord/info_command.h"
+#include "coord/index_list_command.h"
 #include "info/global_stats.h"
 #include "fast_float/fast_float_strtod.h"
 #include "aggregate/aggregate_debug.h"
@@ -173,7 +174,7 @@ RedisModuleString *config_default_scorer = NULL;
 static void DEBUG_DistSearchCommandHandler(void* pd);
 /* ======================= DEBUG ONLY DECLARATIONS ======================= */
 
-static inline bool SearchCluster_Ready() {
+bool SearchCluster_Ready() {
   return NumShards != 0;
 }
 
@@ -1295,43 +1296,6 @@ int ConfigCommand(RedisModuleCtx *ctx, RedisModuleString **argv, int argc) {
   }
 
   RedisModule_EndReply(reply);
-  return REDISMODULE_OK;
-}
-
-// Reads the shard's rdbcompression setting into *enabled, false if it could not be
-// read. getRedisConfigBool() is not usable here: it folds a failed read into its
-// default, and a read failure has to stay distinguishable from both values.
-static bool readRdbCompression(RedisModuleCtx *ctx, bool *enabled) {
-  int value = 0;
-  if (RedisModule_ConfigGetBool(ctx, "rdbcompression", &value) != REDISMODULE_OK) {
-    return false;
-  }
-  *enabled = value != 0;
-  return true;
-}
-
-// Shard-side _FT._LIST; WITHCLUSTERSTATE replies the diagnostic payload the reducer consumes.
-static int IndexListInternal(RedisModuleCtx *ctx, RedisModuleString **argv, int argc) {
-  if (argc > 2) {
-    return RedisModule_WrongArity(ctx);
-  }
-  if (argc == 2) {
-    if (!RMUtil_StringEqualsCaseC(argv[1], "WITHCLUSTERSTATE")) {
-      return RedisModule_ReplyWithError(ctx, QueryError_Strerror(QUERY_ERROR_CODE_ARG_UNRECOGNIZED));
-    }
-
-    bool rdbCompression = false;  // An unreadable setting is reported as incomparable
-    const bool comparable = readRdbCompression(ctx, &rdbCompression);
-    RedisModule_Reply _reply = RedisModule_NewReply(ctx);
-    const char *nodeId = MR_GetLocalNodeId();
-    Indexes_ReplyWithClusterStatePayload(&_reply, nodeId,
-                                         SchemaFingerprint_Recipe(rdbCompression), comparable);
-    MR_ReleaseLocalNodeIdReadLock();
-    return REDISMODULE_OK;
-  }
-
-  RedisModule_Reply _reply = RedisModule_NewReply(ctx);
-  Indexes_List(&_reply, false);
   return REDISMODULE_OK;
 }
 
@@ -3693,11 +3657,11 @@ cleanup:
   return REDISMODULE_OK;
 }
 
-static inline bool cannotBlockCtx(RedisModuleCtx *ctx) {
+bool cannotBlockCtx(RedisModuleCtx *ctx) {
   return RedisModule_GetContextFlags(ctx) & REDISMODULE_CTX_FLAGS_DENY_BLOCKING;
 }
 
-static inline int ReplyBlockDeny(RedisModuleCtx *ctx, const RedisModuleString *cmd) {
+int ReplyBlockDeny(RedisModuleCtx *ctx, const RedisModuleString *cmd) {
   return RMUtil_ReplyWithErrorFmt(ctx, "Cannot perform `%s`: Cannot block", RedisModule_StringPtrLen(cmd, NULL));
 }
 
@@ -4193,449 +4157,6 @@ static int RegisterCoordCursorCommands(RedisModuleCtx* ctx, RedisModuleCommand *
      .setCommandInfo = SetDontCacheInfo, .position = keys}
     };
   return CreateSubCommands(ctx, cursorCommand, subcommands, sizeof(subcommands) / sizeof(SubCommand));
-}
-
-#define FT_LIST_CS_KEY_INDEX       "index"
-#define FT_LIST_CS_KEY_STATUS      "status"
-#define FT_LIST_CS_KEY_WARNING     "warning"
-#define FT_LIST_CS_KEY_MISSING     "missing_from_shards"
-#define FT_LIST_CS_KEY_UNREACHABLE "unreachable_shards"
-#define FT_LIST_CS_STATUS_OK       "ok"
-
-// A shard's comparability gates. Fingerprints are only comparable within one group.
-typedef struct {
-  long long recipe;
-  long long encVer;
-} ClusterStateGates;
-
-// One reported fingerprint, tagged with the gate group that produced it.
-typedef struct {
-  uint32_t group;
-  long long fp;
-} ClusterStateFingerprint;
-
-// Per-index accumulator for IndexListClusterStateReducer.
-typedef struct {
-  arrayof(uint32_t) presentSlots;  // indices into ClusterStateReports.shardIds
-  arrayof(ClusterStateFingerprint) fps;
-  size_t noFingerprint;
-} ClusterStateIndexInfo;
-
-static void ClusterStateIndexInfo_Free(void *p) {
-  ClusterStateIndexInfo *info = p;
-  array_free(info->presentSlots);
-  array_free(info->fps);
-  rm_free(info);
-}
-
-static bool idInArray(const char *id, const char **ids, uint32_t n) {
-  for (uint32_t i = 0; i < n; i++) {
-    if (!strcmp(ids[i], id)) {
-      return true;
-    }
-  }
-  return false;
-}
-
-static bool slotInArray(uint32_t slot, const uint32_t *slots, uint32_t n) {
-  for (uint32_t i = 0; i < n; i++) {
-    if (slots[i] == slot) {
-      return true;
-    }
-  }
-  return false;
-}
-
-// The shard payloads folded into one picture. Strings are borrowed from the
-// replies, which outlive the reducer call.
-typedef struct {
-  TrieMap *byName;  // index name -> ClusterStateIndexInfo
-  // One slot per usable payload: that shard's node id, or "" when it has none
-  // yet. Slots, not ids, are the key, so an unnamed shard still counts.
-  arrayof(const char *) shardIds;
-  size_t nRejected;
-  MRReply *firstError;
-  // Every distinct gate pair the reporting shards sent. More than one means they
-  // disagree on the recipe, so some fingerprints are incomparable. Grouping rather
-  // than holding the shards against one reference matters: with a reference, which
-  // shards got compared depended on which reply arrived first, so a divergence the
-  // replies proved could be dropped when the odd shard happened to answer first.
-  arrayof(ClusterStateGates) gateGroups;
-} ClusterStateReports;
-
-// Folds one shard payload into the picture. An error or malformed reply gets no
-// slot, so it surfaces as non-reporting rather than as divergence.
-static void ClusterStateReports_AddShard(ClusterStateReports *reports, MRReply *r) {
-  if (r && MRReply_Type(r) == MR_REPLY_ERROR) {
-    reports->nRejected++;
-    if (!reports->firstError) {
-      reports->firstError = r;
-    }
-    return;
-  }
-  if (!r || MRReply_Type(r) != MR_REPLY_ARRAY || MRReply_Length(r) != 4) {
-    return;
-  }
-  MRReply *recipe = MRReply_ArrayElement(r, 1);
-  MRReply *encVer = MRReply_ArrayElement(r, 2);
-  MRReply *entries = MRReply_ArrayElement(r, 3);
-  if (MRReply_Type(recipe) != MR_REPLY_INTEGER || MRReply_Type(encVer) != MR_REPLY_INTEGER ||
-      MRReply_Type(entries) != MR_REPLY_ARRAY) {
-    return;
-  }
-  // Validate every entry before taking a slot. Skipping a malformed entry instead
-  // would leave the shard counted as reporting while an index it listed went
-  // unrecorded, and that index would read as missing from a shard that has it -
-  // which is how a future payload with a wider entry would look to this reducer.
-  const size_t nEntries = MRReply_Length(entries);
-  for (size_t j = 0; j < nEntries; j++) {
-    MRReply *pair = MRReply_ArrayElement(entries, j);
-    if (MRReply_Type(pair) != MR_REPLY_ARRAY || MRReply_Length(pair) != 2) {
-      return;
-    }
-  }
-
-  size_t idLen = 0;
-  const char *id = MRReply_String(MRReply_ArrayElement(r, 0), &idLen);
-  array_append(reports->shardIds, id && idLen ? id : "");
-  const uint32_t slot = array_len(reports->shardIds) - 1;
-
-  const ClusterStateGates gates = {.recipe = MRReply_Integer(recipe),
-                                   .encVer = MRReply_Integer(encVer)};
-  uint32_t group = array_len(reports->gateGroups);
-  for (uint32_t g = 0; g < array_len(reports->gateGroups); g++) {
-    if (reports->gateGroups[g].recipe == gates.recipe &&
-        reports->gateGroups[g].encVer == gates.encVer) {
-      group = g;
-      break;
-    }
-  }
-  if (group == array_len(reports->gateGroups)) {
-    reports->gateGroups = array_ensure_append_1(reports->gateGroups, gates);
-  }
-
-  for (size_t j = 0; j < nEntries; j++) {
-    MRReply *pair = MRReply_ArrayElement(entries, j);
-    size_t nameLen = 0;
-    const char *name = MRReply_String(MRReply_ArrayElement(pair, 0), &nameLen);
-    // The payload's strlen-based reply already cut names at their first NUL, and
-    // tm_len_t narrows again; either loss can merge two indexes into one entry.
-    const tm_len_t keyLen = (tm_len_t)nameLen;
-    if (!name || !keyLen) {
-      continue;
-    }
-    ClusterStateIndexInfo *info = TrieMap_Find(reports->byName, name, keyLen);
-    if (info == TRIEMAP_NOTFOUND) {
-      info = rm_calloc(1, sizeof(*info));
-      TrieMap_Add(reports->byName, (char *)name, keyLen, info, NULL);
-    }
-    info->presentSlots = array_ensure_append_1(info->presentSlots, slot);
-    // MRReply_Integer does not check the type and hiredis callocs replies, so a
-    // non-integer would read as 0 and make every such index compare equal.
-    MRReply *fp = MRReply_ArrayElement(pair, 1);
-    if (MRReply_Type(fp) == MR_REPLY_INTEGER) {
-      const ClusterStateFingerprint reported = {.group = group, .fp = MRReply_Integer(fp)};
-      info->fps = array_ensure_append_1(info->fps, reported);
-    } else {
-      info->noFingerprint++;
-    }
-  }
-}
-
-// Shards that reported but did not list this index. *count is how many; the
-// returned array names only those with a node id, so it may be shorter.
-static arrayof(const char *) shardsMissingIndex(const ClusterStateIndexInfo *info,
-                                                const ClusterStateReports *reports, size_t *count) {
-  const uint32_t nSlots = array_len(reports->shardIds);
-  arrayof(const char *) missing = array_new(const char *, nSlots);
-  *count = 0;
-  for (uint32_t slot = 0; slot < nSlots; slot++) {
-    if (slotInArray(slot, info->presentSlots, array_len(info->presentSlots))) {
-      continue;
-    }
-    (*count)++;
-    if (reports->shardIds[slot][0]) {
-      array_append(missing, reports->shardIds[slot]);
-    }
-  }
-  return missing;
-}
-
-// The most distinct schemas any one group of gate-agreeing shards holds for this index.
-// Counting per group is what makes the answer independent of reply order: fingerprints
-// from shards that used different recipes are not comparable, so they are never counted
-// against each other, while a divergence inside any group still surfaces. Which shards
-// are "wrong" is not decidable - there may be no majority - so this is a count only.
-static uint32_t distinctSchemaCount(const ClusterStateIndexInfo *info, uint32_t nGroups) {
-  const uint32_t n = array_len(info->fps);
-  uint32_t worst = 0;
-  for (uint32_t g = 0; g < nGroups; g++) {
-    uint32_t distinct = 0;
-    for (uint32_t i = 0; i < n; i++) {
-      if (info->fps[i].group != g) {
-        continue;
-      }
-      bool seen = false;
-      for (uint32_t j = 0; j < i; j++) {
-        if (info->fps[j].group == g && info->fps[j].fp == info->fps[i].fp) {
-          seen = true;
-          break;
-        }
-      }
-      distinct += !seen;
-    }
-    if (distinct > worst) {
-      worst = distinct;
-    }
-  }
-  return worst;
-}
-
-static void replyShardIds(RedisModule_Reply *reply, const char *key,
-                          arrayof(const char *) ids) {
-  const uint32_t n = array_len(ids);
-  if (n == 0) {
-    return;
-  }
-  RedisModule_ReplyKV_Array(reply, key);
-  for (uint32_t i = 0; i < n; i++) {
-    RedisModule_Reply_SimpleString(reply, ids[i]);
-  }
-  RedisModule_Reply_ArrayEnd(reply);
-}
-
-// Inputs to one index's non-"ok" verdict message.
-typedef struct {
-  size_t nMissing;         // reporting shards that did not list the index
-  uint32_t nSchemas;       // distinct schemas among the comparable reports
-  size_t noFingerprint;
-  size_t reportingShards;
-  size_t nSilent;
-  size_t nRejected;
-  size_t expectedShards;  // shards the fanout asked; the denominator of the two above
-  const char *errorText;
-  bool versionSkew;
-} ClusterStateVerdict;
-
-// Renders the operator-facing warning. Both clauses are appended, not
-// alternatives: a silent shard must not hide a skew between answering shards.
-static void replyClusterStateWarning(RedisModule_Reply *reply, const ClusterStateVerdict *v) {
-  sds msg = sdsnew(INCONSISTENT_INDEX_STATE);
-  const size_t markerLen = sdslen(msg);
-
-  if (v->nMissing > 0 || v->nSchemas > 1) {
-    msg = sdscat(msg, ": ");
-    if (v->nMissing > 0) {
-      msg = sdscatprintf(msg, "index is missing from %zu of %zu reporting shards",
-                     v->nMissing, v->reportingShards);
-    }
-    if (v->nMissing > 0 && v->nSchemas > 1) {
-      msg = sdscat(msg, ", and ");
-    }
-    if (v->nSchemas > 1) {
-      msg = sdscatprintf(msg, "the shards that have it hold %u different schemas",
-                     v->nSchemas);
-    }
-    msg = sdscat(msg, ". Drop the index and recreate it so that all shards agree.");
-  }
-
-  if (v->nSilent > 0 || v->nRejected > 0 || v->versionSkew || v->noFingerprint > 0) {
-    msg = sdscat(msg,
-                   sdslen(msg) > markerLen ? " The rest of the picture cannot be determined: "
-                                   : " cannot be determined: ");
-    const char *sep = "";
-    if (v->nSilent > 0) {
-      msg = sdscatprintf(msg, "%s%zu of %zu shards did not reply", sep, v->nSilent,
-                     v->expectedShards);
-      sep = "; ";
-    }
-    if (v->nRejected > 0) {
-      msg = sdscatprintf(msg, "%s%zu of %zu shards rejected the request", sep,
-                     v->nRejected, v->expectedShards);
-      if (v->errorText) {
-        // Safe unescaped because the warning is replied as a bulk string; the length
-        // cap is there to bound the reply, not to sanitise it.
-        msg = sdscatprintf(msg, " (%.200s)", v->errorText);
-      }
-      sep = "; ";
-    }
-    if (v->versionSkew) {
-      msg = sdscatprintf(msg, "%sshards are running incompatible versions or configurations", sep);
-      sep = "; ";
-    }
-    if (v->noFingerprint > 0) {
-      msg = sdscatprintf(msg, "%s%zu of the reporting shards could not compute a schema fingerprint", sep,
-                     v->noFingerprint);
-    }
-    msg = sdscat(msg, ".");
-  }
-
-  RedisModule_ReplyKV_StringBuffer(reply, FT_LIST_CS_KEY_WARNING, msg, sdslen(msg));
-  sdsfree(msg);
-}
-
-// One { index, status } entry of the public FT._LIST WITHCLUSTERSTATE reply, and the
-// only producer of it - the single-shard path and the reducer both come through here,
-// so the two cannot drift. A NULL verdict is the known-consistent case.
-static void replyClusterStateEntry(RedisModule_Reply *reply, const char *name, size_t nameLen,
-                                   const ClusterStateVerdict *verdict,
-                                   arrayof(const char *) missing,
-                                   arrayof(const char *) unreachable) {
-  RedisModule_Reply_Map(reply);
-  RedisModule_ReplyKV_StringBuffer(reply, FT_LIST_CS_KEY_INDEX, name, nameLen);
-  if (!verdict) {
-    RedisModule_ReplyKV_SimpleString(reply, FT_LIST_CS_KEY_STATUS, FT_LIST_CS_STATUS_OK);
-  } else {
-    RedisModule_ReplyKV_Map(reply, FT_LIST_CS_KEY_STATUS);
-    replyClusterStateWarning(reply, verdict);
-    replyShardIds(reply, FT_LIST_CS_KEY_MISSING, missing);
-    replyShardIds(reply, FT_LIST_CS_KEY_UNREACHABLE, unreachable);
-    RedisModule_Reply_MapEnd(reply);
-  }
-  RedisModule_Reply_MapEnd(reply);
-}
-
-// The name is taken in the same lossy C-string form the shard payload sends, so a
-// single-shard reply renders an index exactly as a multi-shard one does.
-static void replySpecStatusOk(IndexSpec *sp, void *ud) {
-  RedisModule_Reply *reply = ud;
-  const char *name = IndexSpec_FormatName(sp, false);
-  replyClusterStateEntry(reply, name, strlen(name), NULL, NULL, NULL);
-}
-
-// Reducer for FT._LIST WITHCLUSTERSTATE: one map per index across the shards'
-// lists. Divergence the replies prove is reported even when shards are silent.
-static int IndexListClusterStateReducer(struct MRCtx *mc, int count, MRReply **replies) {
-  RedisModuleCtx *ctx = MRCtx_GetRedisCtx(mc);
-  ClusterStateReports reports = {
-      .byName = NewTrieMap(),
-      .shardIds = array_new(const char *, count),
-  };
-  for (int i = 0; i < count; i++) {
-    ClusterStateReports_AddShard(&reports, replies[i]);
-  }
-
-  // An empty array would read as "this cluster has no indexes", so reply an error
-  // instead: the shards' own when one arrived, a fanout failure otherwise.
-  const size_t nReporting = array_len(reports.shardIds);
-  if (nReporting == 0) {
-    const char *err = reports.firstError
-                          ? MRReply_String(reports.firstError, NULL)
-                          : QueryError_Strerror(QUERY_ERROR_CODE_CLUSTER_NO_RESPONSES);
-    TrieMap_Free(reports.byName, ClusterStateIndexInfo_Free);
-    array_free(reports.shardIds);
-    array_free(reports.gateGroups);
-    return RedisModule_ReplyWithError(ctx, err);
-  }
-
-  // The shards the fanout asked, not the current topology: one that joined since would
-  // otherwise be reported as not having replied.
-  size_t expectedCount = 0;
-  const char **expectedIds = MRCtx_GetShardNodeIds(mc, &expectedCount);
-  arrayof(const char *) unreachableIds = array_new(const char *, expectedCount);
-  for (size_t i = 0; i < expectedCount; i++) {
-    if (!idInArray(expectedIds[i], reports.shardIds, array_len(reports.shardIds))) {
-      array_append(unreachableIds, expectedIds[i]);
-    }
-  }
-
-  // Shards the fanout asked that no usable payload came back from. A rejecting shard
-  // is in here too: it was asked, and its error reply took no slot, so its id went
-  // unmatched. That is what keeps nSilent below from underflowing - the count can never
-  // be lower than the rejection tally, since each reply crosses off at most one id.
-  const size_t nNotReporting = array_len(unreachableIds);
-  // A shard that rejected the command is alive, so it is reported as a rejection
-  // rather than counted among those that never answered.
-  const size_t nSilent = nNotReporting - reports.nRejected;
-  // More than one gate group means the reporting shards disagree on the recipe, so some
-  // of their fingerprints could not be compared with each other at all.
-  const bool versionSkew = array_len(reports.gateGroups) > 1;
-  const bool uncertain = nNotReporting > 0 || versionSkew;
-
-  RedisModule_Reply _reply = RedisModule_NewReply(ctx), *reply = &_reply;
-  RedisModule_Reply_Array(reply);
-
-  TrieMapIterator *it = TrieMap_Iterate(reports.byName);
-  char *name;
-  tm_len_t nameLen;
-  void *ptr;
-  while (TrieMapIterator_Next(it, &name, &nameLen, &ptr)) {
-    ClusterStateIndexInfo *info = ptr;
-    size_t nMissing = 0;
-    arrayof(const char *) missing = shardsMissingIndex(info, &reports, &nMissing);
-    const uint32_t nSchemas = distinctSchemaCount(info, array_len(reports.gateGroups));
-    const bool inconsistent = nMissing > 0 || nSchemas > 1;
-
-    if (inconsistent || uncertain || info->noFingerprint > 0) {
-      const ClusterStateVerdict verdict = {
-          .nMissing = nMissing,
-          .nSchemas = nSchemas,
-          .noFingerprint = info->noFingerprint,
-          .reportingShards = nReporting,
-          .nSilent = nSilent,
-          .nRejected = reports.nRejected,
-          .expectedShards = expectedCount,
-          .errorText = reports.firstError ? MRReply_String(reports.firstError, NULL) : NULL,
-          .versionSkew = versionSkew,
-      };
-      replyClusterStateEntry(reply, name, nameLen, &verdict, missing, unreachableIds);
-    } else {
-      replyClusterStateEntry(reply, name, nameLen, NULL, NULL, NULL);
-    }
-
-    array_free(missing);
-  }
-  TrieMapIterator_Free(it);
-
-  RedisModule_Reply_ArrayEnd(reply);
-  RedisModule_EndReply(reply);
-
-  TrieMap_Free(reports.byName, ClusterStateIndexInfo_Free);
-  array_free(reports.shardIds);
-  array_free(reports.gateGroups);
-  array_free(unreachableIds);
-  return REDISMODULE_OK;
-}
-
-// FT._LIST on the coordinator. The no-token form precedes every cluster and
-// blocking check: it must keep working inside MULTI/Lua and when the cluster is down.
-int IndexListCommandHandler(RedisModuleCtx *ctx, RedisModuleString **argv, int argc) {
-  if (argc > 2) {
-    return RedisModule_WrongArity(ctx);
-  }
-  if (argc == 2 && !RMUtil_StringEqualsCaseC(argv[1], "WITHCLUSTERSTATE")) {
-    return RedisModule_ReplyWithError(ctx, QueryError_Strerror(QUERY_ERROR_CODE_ARG_UNRECOGNIZED));
-  }
-  if (argc == 1) {
-    RedisModule_Reply _reply = RedisModule_NewReply(ctx);
-    Indexes_List(&_reply, false);
-    return REDISMODULE_OK;
-  }
-  if (!SearchCluster_Ready()) {
-    return RedisModule_ReplyWithError(ctx, CLUSTERDOWN_ERR);
-  }
-  RS_AutoMemory(ctx);
-
-  if (NumShards == 1) {
-    // Nothing to disagree with, so every local index is trivially consistent.
-    RedisModule_Reply _reply = RedisModule_NewReply(ctx), *reply = &_reply;
-    RedisModule_Reply_Array(reply);
-    Indexes_ForEachSpec(replySpecStatusOk, reply);
-    RedisModule_Reply_ArrayEnd(reply);
-    RedisModule_EndReply(reply);
-    return REDISMODULE_OK;
-  } else if (cannotBlockCtx(ctx)) {
-    return ReplyBlockDeny(ctx, argv[0]);
-  }
-
-  MRCommand cmd = MR_NewCommandFromRedisStrings(argc, argv);
-  MRCommand_SetProtocol(&cmd, ctx);
-  MRCommand_SetPrefix(&cmd, "_FT");
-  struct MRCtx *mrctx = MR_CreateCtx(ctx, 0, NULL, NumShards);
-  // The reducer names the shards that did not reply, so it needs the ones asked.
-  MRCtx_CaptureShardNodeIds(mrctx);
-  MR_Fanout(mrctx, IndexListClusterStateReducer, cmd, true);
-  return REDISMODULE_OK;
 }
 
 int TagValsCommandHandler(RedisModuleCtx *ctx, RedisModuleString **argv, int argc) {
